[#32] refactor : Remove unnecessary keywords from filenames
	파일이름에 꼭 필요하지 않는 키워드를 삭제했습니다
